+++ src/components/common/SnsAuthButton.tsx
